chore: replace separate store/* files with autodux store.js
