+++ ui/src/components/Web3Status.jsx
@@ -5,7 +5,7 @@ import { Button, Chip } from '@material-ui/core';
 import PersonIcon from '@material-ui/icons/Person';
 
 import { useApplicationContext } from '../contexts/Application';
-import { activateConnection, deactivateConnection } from '../store/actions';
+import { activateConnection, deactivateConnection } from '../store';
 
 const useStyles = makeStyles(theme => ({
   divider: {
