--- ui/src/store/types.js
@@ -1,18 +1,0 @@
-export const ACTIVATE_CONNECTION = 'ACTIVATE_CONNECTION';
-export const DEACTIVATE_CONNECTION = 'DEACTIVATE_CONNECTION';
-
-export const SERVER_CONNECTED = 'SERVER_CONNECTED';
-export const SERVER_DISCONNECTED = 'SERVER_DISCONNECTED';
-
-export const UPDATE_PURSES = 'UPDATE_PURSES';
-export const UPDATE_EXCHANGE_AMOUNT = 'UPDATE_EXCHANGE_AMOUNT';
-
-export const CHANGE_PURSE = 'CHANGE_PURSE';
-export const CHANGE_AMOUNT = 'CHANGE_AMOUNT';
-export const SWAP_INPUTS = 'SWAP_INPUTS';
-
-export const CREATE_OFFER = 'CREATE_OFFER';
-
-export const RESET_STATE = 'RESET_STATE';
-
-export const UPDATE_INVITATION_DEPOSIT_ID = 'UPDATE_INVITATION_DEPOSIT_ID';
